fix: Improve bundle size by moving span status
