+++ packages/apm/src/index.ts
@@ -4,5 +4,7 @@ import * as ApmIntegrations from './integrations';
 export { ApmIntegrations as Integrations };
 export { Span, TRACEPARENT_REGEXP } from './span';
 
+export { SpanStatus } from './spanstatus';
+
 // We are patching the global object with our hub extension methods
 addExtensionMethods();
