+++ packages/types/src/index.ts
@@ -17,7 +17,7 @@ export { Runtime } from './runtime';
 export { Scope } from './scope';
 export { SdkInfo } from './sdkinfo';
 export { Severity } from './severity';
-export { Span, SpanContext, SpanStatus } from './span';
+export { Span, SpanContext } from './span';
 export { StackFrame } from './stackframe';
 export { Stacktrace } from './stacktrace';
 export { Status } from './status';
